+++ src/web/assets/base/SproutFormsBaseAsset.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * @link      https://sprout.barrelstrengthdesign.com/
+ * @copyright Copyright (c) Barrel Strength Design LLC
+ * @license   http://sprout.barrelstrengthdesign.com/license
+ */
+
+namespace barrelstrength\sproutforms\web\assets\base;
+
+use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
+
+class SproutFormsBaseAsset extends AssetBundle
+{
+    public function init()
+    {
+        // define the path that your publishable resources live
+        $this->sourcePath = '@barrelstrength/sproutforms/web/assets/base/dist';
+
+        // define the dependencies
+        $this->depends = [
+            CpAsset::class,
+        ];
+
+        // define the relative path to CSS/JS files that should be registered with the page
+        // when this asset bundle is registered
+        $this->css = [
+            'css/sproutforms.css'
+        ];
+
+        parent::init();
+    }
+}