+++ src/web/assets/charts/ChartsAsset.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * @link      https://sprout.barrelstrengthdesign.com/
+ * @copyright Copyright (c) Barrel Strength Design LLC
+ * @license   http://sprout.barrelstrengthdesign.com/license
+ */
+
+namespace barrelstrength\sproutforms\web\assets\charts;
+
+use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
+
+class ChartsAsset extends AssetBundle
+{
+    public function init()
+    {
+        // define the path that your publishable resources live
+        $this->sourcePath = '@barrelstrength/sproutforms/web/assets/charts/dist';
+
+        // define the dependencies
+        $this->depends = [
+            CpAsset::class,
+        ];
+
+        $this->css = [
+            'css/charts-explorer.css'
+        ];
+
+        parent::init();
+    }
+}